pyproject.toml
[project]
name = "job-hunter"
version = "0.1.0"
description = "Findy転職サイトからいいねをくれた企業情報を収集するツール"
dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "openai>=1.75.0",
    "notion-client>=2.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
# パッケージディレクトリを指定
include = [
    "/findy_scraper",
    "/notion_updater",
] 
